// apps/admin/src/shared/api/generated/ja-korea/schemas/financeUpdateItem.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */
import type { FinanceUpdateItemExpenseCategory } from './financeUpdateItemExpenseCategory';

export interface FinanceUpdateItem {
  id: number;
  expenseCategory?: FinanceUpdateItemExpenseCategory;
  /**
     * @minLength 0
     * @maxLength 300
     */
  itemName: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  ratio: number;
  /** @minimum 0 */
  amount: number;
  /** @minimum 1 */
  displayOrder: number;
  version: number;
}